+++ src/fe-common/core/chat-completion.c
@@ -87,7 +87,7 @@ static void last_msg_add(GSList **list, const char *nick, int own, int max)
 		rec = g_new(LAST_MSG_REC, 1);
 		rec->nick = g_strdup(nick);
 
-		if (g_slist_length(*list) == max) {
+		if ((int)g_slist_length(*list) == max) {
 			*list = g_slist_remove(*list,
 					       g_slist_last(*list)->data);
 		}
